Moves async_css code to a new DOM class.
